style: rename files to kebab-case
